+++ CHANGELOG.markdown
@@ -10,6 +10,7 @@
 * Add `Semigroupoids.Do` module with overloads for use with `QualifiedDo`.
 * Add `Apply`, `Alt`, `Plus`, `Bind` and `BindTrans` instances for the CPS
   versions of `WriterT` and `RWST`.
+* Add `psum` function to `Data.Functor.Plus`.
 
 5.3.5 [2020.12.31]
 ------------------