+++ CHANGELOG.markdown
@@ -11,6 +11,7 @@
 * Add `Apply`, `Alt`, `Plus`, `Bind` and `BindTrans` instances for the CPS
   versions of `WriterT` and `RWST`.
 * Add `psum` function to `Data.Functor.Plus`.
+* Add `Categorical` data type.
 
 5.3.5 [2020.12.31]
 ------------------